Updating logs for JobType enum and updated the data type for constant for MultipleInputFileReader and MultipleFileContentDTO classes

+++ src/main/java/edu/charlotte/parser/dto/MultipleFileContentDTO.java
@@ -1,3 +1,3 @@
 package edu.charlotte.parser.dto;
 
-public record MultipleFileContentDTO(String conditionsInputFileContent, String firstFileContent, String secondFileContent, int constantValue) {}
+public record MultipleFileContentDTO(String conditionsInputFileContent, String firstFileContent, String secondFileContent, float constantValue) {}
